Project/match-reports/server/requirements.txt
SQLAlchemy>=2.0
alembic>=1.13
python-dotenv>=1.0
requests>=2.32
pandas>=2.2